+++ src/main/kotlin/com/zepben/evolve/services/network/tracing/feeder/SetDirection.kt
@@ -10,6 +10,7 @@ package com.zepben.evolve.services.network.tracing.feeder
 
 import com.zepben.evolve.cim.iec61970.base.core.Feeder
 import com.zepben.evolve.cim.iec61970.base.core.Terminal
+import com.zepben.evolve.cim.iec61970.base.wires.PowerTransformer
 import com.zepben.evolve.services.network.NetworkService
 import com.zepben.evolve.services.network.tracing.OpenTest
 import com.zepben.evolve.services.network.tracing.traversals.BasicTracker
@@ -102,6 +103,9 @@ class SetDirection {
                 .any { it.normalHeadTerminal == terminal }
         } ?: false
 
+    private fun reachedSubstationTransformer(terminal: Terminal): Boolean =
+        terminal.conductingEquipment.let { ce -> (ce is PowerTransformer) && ce.substations.isNotEmpty() }
+
     private fun flowUpstreamAndQueueNextStraight(
         traversal: BranchRecursiveTraversal<Terminal>,
         terminal: Terminal,
@@ -141,7 +145,7 @@ class SetDirection {
         if (!direction.add(FeederDirection.UPSTREAM))
             return
 
-        if (isFeederHeadTerminal(terminal))
+        if (isFeederHeadTerminal(terminal) || reachedSubstationTransformer(terminal))
             return
 
         val ce = terminal.conductingEquipment ?: return