+++ src/test/kotlin/com/zepben/evolve/services/network/tracing/feeder/SetDirectionTest.kt
@@ -11,10 +11,7 @@
 
 package com.zepben.evolve.services.network.tracing.feeder
 
-import com.zepben.evolve.cim.iec61970.base.core.ConductingEquipment
-import com.zepben.evolve.cim.iec61970.base.core.Feeder
-import com.zepben.evolve.cim.iec61970.base.core.PhaseCode
-import com.zepben.evolve.cim.iec61970.base.core.Terminal
+import com.zepben.evolve.cim.iec61970.base.core.*
 import com.zepben.evolve.services.network.NetworkService
 import com.zepben.evolve.services.network.testdata.PhaseSwapLoopNetwork
 import com.zepben.evolve.services.network.tracing.feeder.DirectionValidator.validateDirections
@@ -117,6 +114,33 @@ class SetDirectionTest {
         n.getT("j3", 2).validateDirections(NONE)
         n.getT("c4", 1).validateDirections(NONE)
         n.getT("c4", 2).validateDirections(NONE)
+    }
+
+    @Test
+    internal fun `stops at zone transformers incase feeder heads are missing`() {
+        //
+        // 1 b0*21--c1--21 tx2 21--c3--2
+        //
+        // * = feeder start
+        //
+        val n = TestNetworkBuilder()
+            .fromBreaker() // b0
+            .toAcls() // c1
+            .toPowerTransformer { addContainer(Substation()) } // tx2
+            .toAcls() // c3
+            .addFeeder("b0", 2)
+            .build()
+
+        DirectionLogger.trace(n["b0"])
+
+        n.getT("b0", 1).validateDirections(NONE)
+        n.getT("b0", 2).validateDirections(DOWNSTREAM)
+        n.getT("c1", 1).validateDirections(UPSTREAM)
+        n.getT("c1", 2).validateDirections(DOWNSTREAM)
+        n.getT("tx2", 1).validateDirections(UPSTREAM)
+        n.getT("tx2", 2).validateDirections(NONE)
+        n.getT("c3", 1).validateDirections(NONE)
+        n.getT("c3", 2).validateDirections(NONE)
     }
 
     @Test